+++ aprsd/clients/aprsis.py
@@ -1,6 +1,5 @@
 import logging
 import select
-import socket
 import threading
 
 import aprslib
@@ -31,23 +30,6 @@ class Aprsdis(aprslib.IS):
     def stop(self):
         self.thread_stop = True
         LOG.info("Shutdown Aprsdis client.")
-
-    def is_socket_closed(self, sock: socket.socket) -> bool:
-        try:
-            # this will try to read bytes without blocking and also without removing them from buffer (peek only)
-            data = sock.recv(16, socket.MSG_DONTWAIT | socket.MSG_PEEK)
-            if len(data) == 0:
-                return True
-        except BlockingIOError:
-            return False  # socket is open and reading from it would block
-        except ConnectionResetError:
-            return True  # socket was closed for some other reason
-        except Exception:
-            self.logger.exception(
-                "unexpected exception when checking if a socket is closed",
-            )
-            return False
-        return False
 
     @wrapt.synchronized(lock)
     def send(self, msg):
