+++ tox.ini
@@ -2,7 +2,7 @@
 minversion = 2.9.0
 skipdist = True
 skip_missing_interpreters = true
-envlist = pep8,py{38,39}
+envlist = pep8,py{39,310}
 #requires = tox-pipenv
 #           pip==22.0.4
 #           pip-tools==5.4.0
